add fully commented-out gradle.properties file
the contents of this file are claimed to be the future syntax or auto-importing
TypeScript type headers using Dukat.

This hasn't been implemented in Kotlin/JS yet (or is broken as of writing),
but will probably be a better way to import these in future,
as opposed to our custom generateHeaders.sh script.